requirements.txt
alembic==1.13.2
asyncpg==0.29.0
bcrypt==4.2.0
fastapi==0.111.1
greenlet==3.0.3
gunicorn==22.0.0
isort==5.13.2
passlib==1.7.4
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose==3.3.0
SQLAlchemy==2.0.31
uvicorn==0.30.3
